requirements.txt
flask
flask-sqlalchemy
flask-login
flask-wtf
email_validator
authlib
python-dotenv
requests
